requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
asyncpg==0.29.0
APScheduler==3.10.4
attrs==23.2.0
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
httplib2==0.20.2
matplotlib==3.9.0
more-itertools==8.10.0
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.0
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pyparsing==2.4.7
python-dotenv==1.0.1
pytz==2024.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
timezonefinder==6.5.0
